Fix show address validation on the shop page
